MebeliGergana/MebeliGergana.Web/Controllers/HomeController.cs
using System;
using System.Web.Mvc;
using System.Web.UI.WebControls;

namespace MebeliGergana.Web.Controllers
{
    public class HomeController : Controller
    {
        
        public ActionResult Index()
        {
            return View();
        }

        public ActionResult About()
        {
            ViewBag.Message = "Мебели Гергана.";

            return View();
        }

        public ActionResult Contact()
        {
            ViewBag.Message = "Your contact page.";

            return View();
        }
    }
}